fix: move integration tests to correct directory (#14)
* fix: move integration tests to correct directory

* refactor: formatting

* refactor: only run unit tests inside GitHub workflow

* feat: add workflow step to run integration tests in the CI separately

* tests: refactor test names for better understandability

* refactor: remove latest commit as it is unnecessary

---------
