+++ .github/workflows/ci.yaml
@@ -31,5 +31,7 @@ jobs:
       run: cargo check
     - name: Linting
       run: cargo clippy
-    - name: Test
-      run: cargo test
+    - name: Unit Tests
+      run: cargo test --lib --color always
+    - name: Integration Tests
+      run: cargo test --test '*' --color always
