+++ sec/src/lib/sec_state_machine/ingestion/retrieval/mod.rs
@@ -160,18 +160,6 @@ mod tests {
     }
 
     #[test]
-    fn should_return_true_when_state_has_computed_the_output() {
-        let mut retrieval_state = Retrieval::default();
-
-        let expected_result = true;
-
-        retrieval_state.compute_output_data();
-        let result = retrieval_state.has_output_data_been_computed();
-
-        assert_eq!(result, expected_result);
-    }
-
-    #[test]
     fn should_return_default_context_data_when_in_initial_state() {
         let retrieval_state = Retrieval::default();
 
@@ -275,18 +263,6 @@ mod tests {
         let expected_result = retrieval_state.get_context_data();
 
         let result = ref_to_retrieval_state.get_context_data();
-
-        assert_eq!(result, expected_result);
-    }
-
-    #[test]
-    fn should_return_true_when_reference_state_has_computed_the_output() {
-        let ref_to_retrieval_state = &mut Retrieval::default();
-
-        let expected_result = true;
-
-        ref_to_retrieval_state.compute_output_data();
-        let result = ref_to_retrieval_state.has_output_data_been_computed();
 
         assert_eq!(result, expected_result);
     }
